Release 01.23.00
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1953 98333e67-4a24-44d7-a75c-e53540dd3050

+++ advanced/src/cpp/server_cgi.cpp
@@ -241,7 +241,7 @@ writeNormalHttpResp(FILE * const  fileP,
         fprintf(fileP, "Set-Cookie: auth=%s\n", authCookie.c_str());
 
     fprintf(fileP, "Content-type: text/xml; charset=\"utf-8\"\n");
-    fprintf(fileP, "Content-length: %u\n", httpBody.size());
+    fprintf(fileP, "Content-length: %u\n", (unsigned)httpBody.size());
     fprintf(fileP, "\n");
 
     // HTTP body
